geting all serviced and editing services

+++ wsapi/src/main/kotlin/com/workshop/wsapi/controllers/ServicesController.kt
@@ -1,7 +1,7 @@
 package com.workshop.wsapi.controllers
 
 import com.workshop.wsapi.models.Service
-import com.workshop.wsapi.models.Visit
+import com.workshop.wsapi.models.ServiceDto
 import com.workshop.wsapi.security.isAdmin
 import com.workshop.wsapi.services.ServiceService
 import org.springframework.beans.factory.annotation.Autowired
@@ -15,10 +15,15 @@ import org.springframework.web.bind.annotation.*
 
 @RestController
 @RequestMapping("/api/services")
-class ServicesController {
+class ServicesController<Optional> {
 
     @Autowired
     private lateinit var serviceService: ServiceService
+
+    @GetMapping("")
+    fun getServices(): ResponseEntity<Any> {
+        return serviceService.getServices()
+    }
 
     @GetMapping("/{id}")
     fun getServiceById(@PathVariable id: Int): String {
@@ -26,8 +31,19 @@ class ServicesController {
     }
 
     @PutMapping("/{id}")
-    fun editService(@PathVariable id: Int, @RequestBody @Validated service: Visit): String {
-        return "Edited service for user $id"
+    fun editService(
+        @PathVariable id: Long,
+        @RequestBody @Validated service: ServiceDto,
+        @AuthenticationPrincipal userDetails: UserDetails
+    ): ResponseEntity<Any> {
+        if (!userDetails.isAdmin()) {
+            return ResponseEntity
+                .status(HttpStatus.FORBIDDEN)
+                .body("You can only access this resource from an Admin Account")
+        }
+
+
+        return serviceService.editService(id, service)
     }
 
     @PostMapping("")
@@ -45,7 +61,12 @@ class ServicesController {
 
 
     @DeleteMapping("/{id}")
-    fun deleteService(@PathVariable id: Long): ResponseEntity<Any> {
+    fun deleteService(@PathVariable id: Long, @AuthenticationPrincipal userDetails: UserDetails): ResponseEntity<Any> {
+        if (!userDetails.isAdmin()) {
+            return ResponseEntity
+                .status(HttpStatus.FORBIDDEN)
+                .body("You can only access this resource from an Admin Account")
+        }
         return serviceService.deleteService(id)
     }
 }
